Rename logo folder
Renaming the /logo folder to /images to make it more general. Next PR will add more icons to this folder.
