+++ packages/renderer/src/lib/images/Logo.svelte
